Set XDG_CACHE_HOME and GOFLAGS for Go projects
That should fix the behavior in which the default GOCACHE isn't writable
in pre-commit.ci, because GOCACHE defaults to $HOME/.cache/go-build on
Linux, and $HOME ins't writable. Setting it to /tmp is kind of a dirty
trick, but I figure pre-commit doesn't really care about GOCACHE since
it caches the $GOPATH anyways? (Alternatively we can have pre-commit
cache $GOCACHE too).

This also sets GOFLAGS to modify Go's default behavior when fetching
modules locally: pre-commit run a rmtree at the end of the process, and
Go's default behavior is to create the modcache in read-only mode, which
causes rmtree to fail. By setting -modcacherw, we tell Go to create
modcache in read-write mode.